Create part-time job
Snoopy lost his wallet, but he is afraid to tell his family, so he decided to do a part-time job to earn money in weekends. He is selling
washing machines in a big market, he will get p dollars by sold a washing machine, and the manager promised him that snoopy will get
at least m dollars a day, it means that if snoopy didn't sold as many washing machines to get m dollars, the manager will pay him m dollars also.


Now we told you the p and m, and the number of washing machines snoopy sold in a day, which we call it n, help snoopy to calculate
how many dollars he can get.

输入格式
There are multiple test cases in the input file. Each line is a test case, there is three integers m, p, and n which are descript in the problem. Input is terminated by p = n = m = 0. ( 1 <= p, n, m <= 100 )

输出格式
For each test case, output an integer that the payment snoopy can get in a single line.

样例输入
20 5 3
20 5 5
0 0 0
样例输出
20
25